+++ client/app/pages/dashboards/public-dashboard-page.js
@@ -2,7 +2,6 @@ import PromiseRejectionError from '@/lib/promise-rejection-error';
 import logoUrl from '@/assets/images/redash_icon_small.png';
 import template from './public-dashboard-page.html';
 import dashboardGridOptions from '@/config/dashboard-grid-options';
-import _ from 'lodash';
 import './dashboard.less';
 
 function loadDashboard($http, $route) {
@@ -19,7 +18,6 @@ const PublicDashboardPage = {
     'ngInject';
 
     this.filters = [];
-    let currentQueryResultsErrorData;
 
     this.dashboardGridOptions = Object.assign({}, dashboardGridOptions, {
       resizable: { enabled: false },
@@ -72,38 +70,6 @@ const PublicDashboardPage = {
     };
 
     this.refreshDashboard();
-
-    this.getQueryResultsErrorData = () => {
-      const dashboardErrors = _.map(this.dashboard.widgets, (widget) => {
-        // get result
-        const result = widget.getQueryResult();
-        if (!result) {
-          return null;
-        }
-
-        // get error data
-        const errorData = result.getErrorData();
-        if (_.isEmpty(errorData)) {
-          return null;
-        }
-
-        // dashboard params only
-        const localParamNames = _.map(widget.getLocalParameters(), p => p.name);
-        const filtered = _.omit(errorData.parameters, localParamNames);
-
-        return filtered;
-      });
-
-      const merged = _.assign({}, ...dashboardErrors);
-      const errorData = _.isEmpty(merged) ? null : { parameters: merged };
-
-      // avoiding Angular infdig (ANGULAR_REMOVE_ME)
-      if (!_.isEqual(currentQueryResultsErrorData, errorData)) {
-        currentQueryResultsErrorData = errorData;
-      }
-
-      return currentQueryResultsErrorData;
-    };
   },
 };
